mon: Monitor: the monitor as a generic k/v store
Building up on the Single-Paxos and our existing k/v store that backs
the monitor, we now introduce a simple mechanism that makes the
monitors to act as a generic k/v store available to the cluster.

Users can put, get, delete, list and check for values using the
following commands:

 - ceph store put <key> <value>
  or
 - ceph store put <key> -i <in-file>

 - ceph store get <key>
  or
 - ceph store get <key> -o <out-file>

 - ceph store delete <key>

 - ceph store list

 - ceph store exists <key>

NOTE: this is just a prototype; the final version should be fully
contained in a service of its own, instead of being completely coded on
Monitor.cc.

Signed-off-by: Joao Eduardo Luis <joao.luis@inktank.com>